Use consistent types to keep the number of warnings down Add restrict and const where appropriate

+++ src/MVAnalyse.c
@@ -115,8 +115,8 @@ static const VSFrame *VS_CC mvanalyseGetFrame(int n, int activationReason, void 
 
         const uint8_t *pSrc[3] = { NULL };
         const uint8_t *pRef[3] = { NULL };
-        int nSrcPitch[3] = { 0 };
-        int nRefPitch[3] = { 0 };
+        ptrdiff_t nSrcPitch[3] = { 0 };
+        ptrdiff_t nRefPitch[3] = { 0 };
 
         int nref;
 
@@ -519,7 +519,7 @@ static void VS_CC mvanalyseCreate(const VSMap *in, VSMap *out, void *userData, V
 #define ERROR_SIZE 1024
     char errorMsg[ERROR_SIZE] = "Analyse: failed to retrieve first frame from super clip. Error message: ";
     size_t errorLen = strlen(errorMsg);
-    const VSFrame *evil = vsapi->getFrame(0, d.node, errorMsg + errorLen, ERROR_SIZE - errorLen);
+    const VSFrame *evil = vsapi->getFrame(0, d.node, errorMsg + errorLen, ERROR_SIZE - (int)errorLen);
 #undef ERROR_SIZE
     if (!evil) {
         vsapi->mapSetError(out, errorMsg);
